refactor: extract schema-directed payload traversal
Move canonical extraction, annotation resolution, structural-field, and
declared-path traversal into one cycle-safe leaf module. Keep migration
and render orchestration in _runtime.

All 53 definitions are source- and AST-identical to origin/main.
Preserve the 0.3.0 exclude and exclude_if wire boundary verbatim.

Validation: 858 tests passed on Pydantic 2.13.4 (95.85% coverage).
Validation: 858 tests passed on Pydantic 2.12.3 (95.76% coverage).
Validation: 502 focused tests and 22 independent exclusions passed.
Validation: Ruff, ty, strict mypy, Vulture, and strict docs passed.
Validation: immutable v0.3.0 contract passed.

Closes #112
Refs #94
